API serializer proposal
