+++ sandbox/Example.Core.SubscribeQueueGroup/Program.cs
@@ -1,5 +1,4 @@
 // > nats pub foo.xyz --count=10 "my_message_{{ Count }}"
-using System.Text;
 using Microsoft.Extensions.Logging;
 using NATS.Client.Core;
 
@@ -12,13 +11,12 @@ Print("[1][CON] Connecting...\n");
 await using var connection1 = new NatsConnection(options);
 
 Print($"[1][SUB] Subscribing to subject '{subject}'...\n");
-var sub1 = await connection1.SubscribeAsync(subject, queueGroup: "My-Workers");
+var sub1 = await connection1.SubscribeAsync<string>(subject, queueGroup: "My-Workers");
 var task1 = Task.Run(async () =>
 {
     await foreach (var msg in sub1.Msgs.ReadAllAsync())
     {
-        var data = Encoding.UTF8.GetString(msg.Data.ToArray());
-        Print($"[1][RCV] {msg.Subject}: {data}\n");
+        Print($"[1][RCV] {msg.Subject}: {msg.Data}\n");
     }
 });
 
@@ -28,13 +26,12 @@ Print("[2][CON] Connecting...\n");
 await using var connection2 = new NatsConnection(options);
 
 Print($"[2][SUB] Subscribing to subject '{subject}'...\n");
-var sub2 = await connection2.SubscribeAsync(subject, queueGroup: "My-Workers");
+var sub2 = await connection2.SubscribeAsync<string>(subject, queueGroup: "My-Workers");
 var task2 = Task.Run(async () =>
 {
     await foreach (var msg in sub2.Msgs.ReadAllAsync())
     {
-        var data = Encoding.UTF8.GetString(msg.Data.ToArray());
-        Print($"[2][RCV] {msg.Subject}: {data}\n");
+        Print($"[2][RCV] {msg.Subject}: {msg.Data}\n");
     }
 });
