+++ src/NATS.Client.Core/NatsConnection.RequestReply.cs
@@ -1,4 +1,3 @@
-using System.Buffers;
 using System.Runtime.CompilerServices;
 
 namespace NATS.Client.Core;
@@ -36,30 +35,6 @@ public partial class NatsConnection
     }
 
     /// <inheritdoc />
-    public async ValueTask<NatsMsg?> RequestAsync(
-        string subject,
-        ReadOnlySequence<byte> payload = default,
-        NatsHeaders? headers = default,
-        NatsPubOpts? requestOpts = default,
-        NatsSubOpts? replyOpts = default,
-        CancellationToken cancellationToken = default)
-    {
-        var opts = SetReplyOptsDefaults(replyOpts);
-
-        await using var sub = await RequestSubAsync(subject, payload, headers, requestOpts, opts, cancellationToken).ConfigureAwait(false);
-
-        if (await sub.Msgs.WaitToReadAsync(cancellationToken).ConfigureAwait(false))
-        {
-            if (sub.Msgs.TryRead(out var msg))
-            {
-                return msg;
-            }
-        }
-
-        return null;
-    }
-
-    /// <inheritdoc />
     public async IAsyncEnumerable<NatsMsg<TReply?>> RequestManyAsync<TRequest, TReply>(
         string subject,
         TRequest? data,
@@ -77,32 +52,6 @@ public partial class NatsConnection
             {
                 // Received end of stream sentinel
                 if (msg.Data is null)
-                {
-                    yield break;
-                }
-
-                yield return msg;
-            }
-        }
-    }
-
-    /// <inheritdoc />
-    public async IAsyncEnumerable<NatsMsg> RequestManyAsync(
-        string subject,
-        ReadOnlySequence<byte> payload = default,
-        NatsHeaders? headers = default,
-        NatsPubOpts? requestOpts = default,
-        NatsSubOpts? replyOpts = default,
-        [EnumeratorCancellation] CancellationToken cancellationToken = default)
-    {
-        await using var sub = await RequestSubAsync(subject, payload, headers, requestOpts, replyOpts, cancellationToken).ConfigureAwait(false);
-
-        while (await sub.Msgs.WaitToReadAsync(cancellationToken).ConfigureAwait(false))
-        {
-            while (sub.Msgs.TryRead(out var msg))
-            {
-                // Received end of stream sentinel
-                if (msg.Data.Length == 0)
                 {
                     yield break;
                 }
