+++ src/NATS.Client.Core/NatsRequestExtensions.cs
@@ -1,5 +1,3 @@
-using System.Buffers;
-
 namespace NATS.Client.Core;
 
 public static class NatsRequestExtensions
@@ -17,7 +15,7 @@ public static class NatsRequestExtensions
     /// <returns>Returns the <see cref="NatsMsg{T}"/> received from the responder as reply.</returns>
     /// <exception cref="OperationCanceledException">Raised when cancellation token is used</exception>
     /// <remarks>
-    /// Response can be (null) or one <see cref="NatsMsg"/>.
+    /// Response can be (null) or one <see cref="NatsMsg{T}"/>.
     /// Reply option's max messages will be set to 1.
     /// if reply option's timeout is not defined then it will be set to NatsOpts.RequestTimeout.
     /// </remarks>
@@ -39,48 +37,13 @@ public static class NatsRequestExtensions
             cancellationToken);
     }
 
-    /// <summary>
-    /// Request and receive a single reply from a responder.
-    /// </summary>
-    /// <param name="nats">NATS connection</param>
-    /// <param name="msg">Message to be sent as request</param>
-    /// <param name="requestOpts">Request publish options</param>
-    /// <param name="replyOpts">Reply handler subscription options</param>
-    /// <param name="cancellationToken">Cancel this request</param>
-    /// <returns>Returns the <see cref="NatsMsg"/> received from the responder as reply.</returns>
-    /// <exception cref="OperationCanceledException">Raised when cancellation token is used</exception>
-    /// <remarks>
-    /// Response can be (null) or one <see cref="NatsMsg"/>.
-    /// Reply option's max messages will be set to 1.
-    /// if reply option's timeout is not defined then it will be set to NatsOpts.RequestTimeout.
-    /// </remarks>
-    public static ValueTask<NatsMsg?> RequestAsync(
-        this INatsConnection nats,
-        in NatsMsg msg,
-        in NatsPubOpts? requestOpts = default,
-        in NatsSubOpts? replyOpts = default,
-        CancellationToken cancellationToken = default)
-    {
-        CheckMsgForRequestReply(msg);
-
-        return nats.RequestAsync(
-            msg.Subject,
-            payload: new ReadOnlySequence<byte>(msg.Data),
-            msg.Headers,
-            requestOpts,
-            replyOpts,
-            cancellationToken);
-    }
-
-    internal static void CheckMsgForRequestReply(in NatsMsg msg) => CheckForRequestReply(msg.ReplyTo);
-
     internal static void CheckMsgForRequestReply<T>(in NatsMsg<T> msg) => CheckForRequestReply(msg.ReplyTo);
 
     private static void CheckForRequestReply(string? replyTo)
     {
         if (!string.IsNullOrWhiteSpace(replyTo))
         {
-            throw new NatsException($"Can't set {nameof(NatsMsg.ReplyTo)} for a request");
+            throw new NatsException($"Can't set reply-to for a request");
         }
     }
 }
