+++ tests/NATS.Client.Core.Tests/RequestReplyTest.cs
@@ -260,11 +260,6 @@ public class RequestReplyTest
     [Fact]
     public async Task Request_reply_binary_test()
     {
-        static ReadOnlySequence<byte> ToSeq(string input)
-        {
-            return new ReadOnlySequence<byte>(Encoding.ASCII.GetBytes(input));
-        }
-
         static string ToStr(ReadOnlyMemory<byte> input)
         {
             return Encoding.ASCII.GetString(input.Span);
@@ -274,22 +269,22 @@ public class RequestReplyTest
 
         await using var server = NatsServer.Start();
         await using var nats = server.CreateClientConnection();
-        await using var sub = await nats.SubscribeAsync("foo", cancellationToken: cts.Token);
+        await using var sub = await nats.SubscribeAsync<string>("foo", cancellationToken: cts.Token);
         var reg = sub.Register(async m =>
         {
-            if (ToStr(m.Data) == "1")
+            if (m.Data == "1")
             {
-                await m.ReplyAsync(payload: ToSeq("qw"), cancellationToken: cts.Token);
-                await m.ReplyAsync(payload: ToSeq("er"), cancellationToken: cts.Token);
-                await m.ReplyAsync(payload: ToSeq("ty"), cancellationToken: cts.Token);
-                await m.ReplyAsync(payload: default, cancellationToken: cts.Token); // sentinel
+                await m.ReplyAsync("qw", cancellationToken: cts.Token);
+                await m.ReplyAsync("er", cancellationToken: cts.Token);
+                await m.ReplyAsync("ty", cancellationToken: cts.Token);
+                await m.ReplyAsync(default(string), cancellationToken: cts.Token); // sentinel
             }
         });
 
         var writer = new ArrayBufferWriter<byte>();
-        await foreach (var msg in nats.RequestManyAsync("foo", ToSeq("1"), cancellationToken: cts.Token))
+        await foreach (var msg in nats.RequestManyAsync<string, string>("foo", "1", cancellationToken: cts.Token))
         {
-            writer.Write(msg.Data.Span);
+            writer.Write(Encoding.UTF8.GetBytes(msg.Data!));
         }
 
         var buffer = ToStr(writer.WrittenMemory);
